refactor: cheer/store 패키지 분리
