+++ src/main/java/eatda/controller/cheer/CheerPreviewResponse.java
@@ -1,6 +1,6 @@
-package eatda.controller.store;
+package eatda.controller.cheer;
 
-import eatda.domain.store.Cheer;
+import eatda.domain.cheer.Cheer;
 import eatda.domain.store.Store;
 
 public record CheerPreviewResponse(
