+++ src/test/java/eatda/document/BaseDocumentTest.java
@@ -9,8 +9,8 @@ import eatda.exception.BusinessErrorCode;
 import eatda.exception.EtcErrorCode;
 import eatda.service.article.ArticleService;
 import eatda.service.auth.AuthService;
+import eatda.service.cheer.CheerService;
 import eatda.service.member.MemberService;
-import eatda.service.store.CheerService;
 import eatda.service.store.StoreService;
 import eatda.service.story.StoryService;
 import io.restassured.RestAssured;
