+++ src/test/java/eatda/domain/store/StoreSearchFilterTest.java
@@ -1,4 +1,4 @@
-package eatda.service.store;
+package eatda.domain.store;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,10 +23,10 @@ class StoreSearchFilterTest {
 
         @Test
         void 음식점이_아니거나_서울에_위치하지_않는_가게는_제외한다() {
-            StoreSearchResult store1 = createStore("1", "서울음식점1","FD6", "서울 강남구 대치동 896-33");
-            StoreSearchResult store2 = createStore("2", "카페", "CD2","서울 강남구 대치동 896-33");
-            StoreSearchResult store3 = createStore("3", "서울음식점2", "FD6","서울 강남구 대치동 896-33");
-            StoreSearchResult store4 = createStore("4", "부산음식점", "FD6","부산 연제구 연산동 632-8");
+            StoreSearchResult store1 = createStore("1", "서울음식점1", "FD6", "서울 강남구 대치동 896-33");
+            StoreSearchResult store2 = createStore("2", "카페", "CD2", "서울 강남구 대치동 896-33");
+            StoreSearchResult store3 = createStore("3", "서울음식점2", "FD6", "서울 강남구 대치동 896-33");
+            StoreSearchResult store4 = createStore("4", "부산음식점", "FD6", "부산 연제구 연산동 632-8");
 
             List<StoreSearchResult> searchResults = List.of(store1, store2, store3, store4);
             List<StoreSearchResult> actual = storeSearchFilter.filterSearchedStores(searchResults);
