+++ src/test/java/eatda/fixture/CheerGenerator.java
@@ -1,8 +1,8 @@
 package eatda.fixture;
 
 import eatda.domain.ImageKey;
+import eatda.domain.cheer.Cheer;
 import eatda.domain.member.Member;
-import eatda.domain.store.Cheer;
 import eatda.domain.store.Store;
 import eatda.repository.store.CheerRepository;
 import eatda.util.DomainUtils;
